Move coloring info logic into a dedicated helper
